SOLID principles verified and few methods adjusted.
